Added button data together with event types refactoring
