+++ examples/src/pages/Migration.tsx
@@ -15,7 +15,7 @@ import {
   CameraControlsOptions,
   TreeIndexNodeCollection,
   CogniteModelBase,
-  DefaultCameraManager
+  DefaultCameraManager,
 } from '@cognite/reveal';
 import { DebugCameraTool, DebugLoadedSectorsTool, DebugLoadedSectorsToolOptions, ExplodedViewTool, AxisViewTool } from '@cognite/reveal/tools';
 import * as reveal from '@cognite/reveal';
@@ -227,7 +227,12 @@ export function Migration() {
       debugStatsGui.add(guiState.debug.stats, 'textures').name('Textures');
       debugStatsGui.add(guiState.debug.stats, 'renderTime').name('Ms/frame');
 
-      viewer.on('sceneRendered', sceneRenderedEventArgs => {
+      viewer.on('sceneRendered', (sceneRenderedEventArgs: {
+        frameNumber: number;
+        renderTime: number;
+        renderer: THREE.WebGLRenderer;
+        camera: THREE.PerspectiveCamera;
+      }) => {
         guiState.debug.stats.drawCalls = sceneRenderedEventArgs.renderer.info.render.calls;
         guiState.debug.stats.points = sceneRenderedEventArgs.renderer.info.render.points;
         guiState.debug.stats.triangles = sceneRenderedEventArgs.renderer.info.render.triangles;
@@ -369,8 +374,8 @@ export function Migration() {
 
       const inspectNodeUi = new InspectNodeUI(gui.addFolder('Last clicked node'), client);
 
-      viewer.on('click', async event => {
-        const { offsetX, offsetY } = event;
+      viewer.on('click', async (event: { offsetX: any; offsetY: any; button: any; }) => {
+        const { offsetX, offsetY, button } = event; 
         console.log('2D coordinates', event);
         const intersection = await viewer.getIntersectionFromPixel(offsetX, offsetY);
         if (intersection !== null) {
