+++ viewer/packages/camera-manager/src/CameraManager.ts
@@ -2,7 +2,8 @@
  * Copyright 2022 Cognite AS
  */
 
-import { CameraChangeDelegate, CameraState } from './types';
+import { CameraState } from './types';
+import { CameraChangeDelegate } from '@reveal/utilities';
 
 /**
  * Interface for manager responsible for all manipulations to the camera,
