+++ viewer/packages/camera-manager/src/DefaultCameraManager.ts
@@ -8,13 +8,11 @@ import { ComboControls } from './ComboControls';
 import {
   CameraManagerCallbackData,
   CameraControlsOptions,
-  CameraChangeDelegate,
-  PointerEventDelegate,
   CameraState
 } from './types';
 import { CameraManager } from './CameraManager';
 import { CameraManagerHelper } from './CameraManagerHelper';
-import { assertNever, EventTrigger, InputHandler, disposeOfAllEventListeners } from '@reveal/utilities';
+import { assertNever, EventTrigger, InputHandler, disposeOfAllEventListeners, CameraChangeDelegate, PointerEventDelegate} from '@reveal/utilities';
 
 /**
  * Default implementation of {@link CameraManager}. Uses target-based orbit controls combined with
