+++ viewer/packages/utilities/src/events/index.ts
@@ -4,3 +4,4 @@
 export { clickOrTouchEventOffset } from './clickOrTouchEventOffset';
 export { EventTrigger } from './EventTrigger';
 export { InputHandler, disposeOfAllEventListeners } from './InputHandler';
+export * from './types';
